image/new/requirements.txt
fastapi
psycopg2-binary
uvicorn
pydantic
pandas
